+++ Grains/Grains.csproj
@@ -7,13 +7,9 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.Orleans.CodeGenerator" Version="3.5.1" />
-    <PackageReference Include="Microsoft.Orleans.Core" Version="3.5.1" />
-    <PackageReference Include="Microsoft.Orleans.CodeGenerator.MSBuild" Version="3.5.1">
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-      <PrivateAssets>all</PrivateAssets>
-    </PackageReference>
-    <PackageReference Include="Microsoft.Orleans.Persistence.AdoNet" Version="3.5.1" />
+    <PackageReference Include="Microsoft.Orleans.Core" Version="7.2.6" />
+    <PackageReference Include="Microsoft.Orleans.Persistence.AdoNet" Version="7.2.6" />
+    <PackageReference Include="Microsoft.Orleans.Sdk" Version="7.2.6" />
     <PackageReference Include="Npgsql" Version="5.0.7" />
     <PackageReference Include="Serilog" Version="3.1.1" />
     <PackageReference Include="Serilog.AspNetCore" Version="8.0.1" />
