+++ Shared/Shared.csproj
@@ -7,8 +7,8 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.Orleans.CodeGenerator" Version="3.5.1" />
-    <PackageReference Include="Microsoft.Orleans.Core" Version="3.5.1" />
+    <PackageReference Include="Microsoft.Orleans.Core" Version="7.2.6" />
+    <PackageReference Include="Microsoft.Orleans.Sdk" Version="7.2.6" />
   </ItemGroup>
 
 </Project>
