+++ SiloHost/SiloHost.csproj
@@ -11,14 +11,16 @@
     <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="8.0.0" />
     <PackageReference Include="Microsoft.Extensions.Logging.Console" Version="8.0.0" />
     <PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="8.0.0" />
-    <PackageReference Include="Microsoft.Orleans.Clustering.AdoNet" Version="3.5.1" />
-    <PackageReference Include="Microsoft.Orleans.Core" Version="3.5.1" />
-    <PackageReference Include="Microsoft.Orleans.OrleansRuntime" Version="3.5.1" />
-    <PackageReference Include="Microsoft.Orleans.Persistence.AdoNet" Version="3.5.1" />
-    <PackageReference Include="Microsoft.Orleans.Server" Version="3.5.1" />
+    <PackageReference Include="Microsoft.Orleans.Clustering.AdoNet" Version="7.2.6" />
+    <PackageReference Include="Microsoft.Orleans.Core" Version="7.2.6" />
+    <PackageReference Include="Microsoft.Orleans.Persistence.AdoNet" Version="7.2.6" />
+    <PackageReference Include="Microsoft.Orleans.Runtime" Version="7.2.6" />
+    <PackageReference Include="Microsoft.Orleans.Sdk" Version="7.2.6" />
+    <PackageReference Include="Microsoft.Orleans.Serialization.NewtonsoftJson" Version="7.2.6" />
+    <PackageReference Include="Microsoft.Orleans.Server" Version="7.2.6" />
     <PackageReference Include="MySql.Data" Version="8.0.28" />
     <PackageReference Include="MySqlConnector" Version="1.3.13" />
-    <PackageReference Include="OrleansDashboard" Version="3.5.1" />
+    <PackageReference Include="OrleansDashboard" Version="7.2.2" />
     <PackageReference Include="Serilog" Version="3.1.1" />
     <PackageReference Include="Serilog.Extensions.Logging" Version="8.0.0" />
     <PackageReference Include="Serilog.Sinks.Console" Version="5.0.1" />
