+++ WebApiClient/WebApiClient.csproj
@@ -10,14 +10,12 @@
   <ItemGroup>
     <PackageReference Include="Microsoft.AspNetCore.OpenApi" Version="8.0.1" />
     <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="8.0.0" />
-    <PackageReference Include="Microsoft.Orleans.Client" Version="3.5.1" />
-    <PackageReference Include="Microsoft.Orleans.Core" Version="3.5.1" />
-    <PackageReference Include="Microsoft.Orleans.Server" Version="3.5.1" />
-    <PackageReference Include="Microsoft.Orleans.OrleansProviders" Version="3.5.1" />
-    <PackageReference Include="Microsoft.Orleans.OrleansRuntime" Version="3.5.1" />
+    <PackageReference Include="Microsoft.Orleans.Client" Version="7.2.6" />
+    <PackageReference Include="Microsoft.Orleans.Clustering.AdoNet" Version="7.2.6" />
+    <PackageReference Include="Microsoft.Orleans.Core" Version="7.2.6" />
+    <PackageReference Include="Microsoft.Orleans.Sdk" Version="7.2.6" />
     <PackageReference Include="Microsoft.Orleans.OrleansAWSUtils" Version="3.0.0-beta1" />
     <PackageReference Include="Microsoft.Orleans.OrleansSQLUtils" Version="3.0.0-beta1" />
-    <PackageReference Include="Microsoft.Orleans.OrleansZooKeeperUtils" Version="3.5.1" />
     <PackageReference Include="MySql.Data" Version="8.0.28" />
     <PackageReference Include="MySqlConnector" Version="1.3.13" />
